fix: fix ref parser to support dl.acm exports
moved Reference to proper file
